Refactor EoL rename module and migrate to TreeOfLife toolbox.
Moved the EoL renaming module from DD_tools to TreeOfLife_toolbox, updating imports, paths, and structure. Enhanced documentation and added a README to clarify functionality, components, and usage. No changes were made to the core logic, ensuring backward compatibility.
